Move opengl initilisation into the window module
